book_loader/requirements.txt
pytest==7.2.1
requests==2.31.0
lxml==4.9.0
requests-html==0.10.0
black==23.3.0
pyyaml==6.0.1
firebase-admin==6.5.0
google-cloud-firestore==2.16.1